lookup workflow name for WorkflowProperties
We currently have to hit the db to lookup a workflow name, so we move a
bunch of functions under argoCabooseRunner to get context and a database
connection. But by looking up this name, we can provide much better
error messages when processing callbacks.

Note that the TemplateName field has been removed from
WorkflowProperties--it was argo-specific and didn't belong on that type
anyway.